# Migrate h64 to BigInts, add h32 and h64 streaming APIs, plus performance improvements