--- a/.eslintrc.js
+++ b/.eslintrc.js
@@ -8,6 +8,7 @@ module.exports = {
     }
   },
   env: {
+    es2020: true,
     es6: true,
     browser: true,
     node: true