--- a/src/index.js
+++ b/src/index.js
@@ -115,7 +115,7 @@ async function xxhash() {
   const defaultSeed = 0;
   const defaultBigSeed = 0n;
 
-  function h32String(str, seed = defaultSeed) {
+  function h32(str, seed = defaultSeed) {
     // https://developer.mozilla.org/en-US/docs/Web/API/TextEncoder/encodeInto#buffer_sizing
     // By sizing the buffer to 3 * string-length we guarantee that the buffer
     // will be appropriately sized for the utf-8 encoding of the string.
@@ -125,7 +125,7 @@ async function xxhash() {
     );
   }
 
-  function h64String(str, seed = defaultBigSeed) {
+  function h64(str, seed = defaultBigSeed) {
     growMemory(str.length * 3, 0);
     return forceUnsigned64(
       xxh64(0, encoder.encodeInto(str, memory).written, seed)
@@ -133,10 +133,10 @@ async function xxhash() {
   }
 
   return {
-    h32(str, seed = defaultSeed) {
-      return h32String(str, seed).toString(16).padStart(8, "0");
+    h32,
+    h32ToString(str, seed = defaultSeed) {
+      return h32(str, seed).toString(16).padStart(8, "0");
     },
-    h32String,
     h32Raw(inputBuffer, seed = defaultSeed) {
       growMemory(inputBuffer.byteLength, 0);
       memory.set(inputBuffer);
@@ -152,10 +152,10 @@ async function xxhash() {
         forceUnsigned32
       );
     },
-    h64(str, seed = defaultBigSeed) {
-      return h64String(str, seed).toString(16).padStart(16, "0");
+    h64,
+    h64ToString(str, seed = defaultBigSeed) {
+      return h64(str, seed).toString(16).padStart(16, "0");
     },
-    h64String,
     h64Raw(inputBuffer, seed = defaultBigSeed) {
       growMemory(inputBuffer.byteLength, 0);
       memory.set(inputBuffer);