… package.json
… src/xxhash.wasm