@marcusdarmstrong 2022-01-13
> if you pass a number to the 64-bit API is it still handled correctly and just limits the maximum value or are the bytes misinterpreted by the WASM (as in float in memory is interpreted as a long)?

Passing a number as a BigInt seed is a runtime error: `TypeError: Cannot convert 1 to a BigInt`. My preference would be to avoid the conversion, as constructing a new BigInt is an allocation and as such involves some legitimate overhead.

I've added the relevant testcases, removed the no-longer-compatible CI targets, and updated the readme with 1. a streaming example, 2. a mention of the limitations of the string api and 3. a discussion of engine requirements. After some thought, I've also added two new API methods here—feedback of course welcome—`h32String` and `h64String` that provide raw numeric hashes with the performance of the encodeInto API. Leaving an ergonomic default but providing access to the best possible performance case seemed like the best option to me... There's unfortunately no way to emulate the encodeInto performance via the Raw methods.