@jungomi 2022-01-13
> > if you pass a number to the 64-bit API is it still handled correctly and just limits the maximum value or are the bytes misinterpreted by the WASM (as in float in memory is interpreted as a long)?
> 
> Passing a number as a BigInt seed is a runtime error: `TypeError: Cannot convert 1 to a BigInt`. My preference would be to avoid the conversion, as constructing a new BigInt is an allocation and as such involves some legitimate overhead.

Ah okay, then it's fine as it is.

The `h32String`/`h64String` methods are definitely a nice solution, but the names are counter-intuitive, because it would kind of be expected to be the other way around. That brings up the question, whether the default methods (`h32`/`h64`) should actually be returning strings or whether the string version is an additional one. In terms of API design it's certainly more elegant to have the default returning numbers and have a convenience method to directly convert it to a string.
If users expect a string output from the default methods (probably mostly due to backwards-compatibility), I think the `h**String` should be renamed to something that better indicates the difference being it returning a number instead of a string.
I'm personally in favour of having the default methods return a number and adding the convenience methods for string outputs, but the only thing I'm worried about if these changes will get caught, because if the TypeScript types are not checked, it could lead to silent misbehaviour (most likely with string concatenation), because there will not be any runtime error from that.
There are other breaking changes and it's a new major version, so I guess it doesn't really matter. What do you think?

Otherwise the rest looks fine to me.