@marcusdarmstrong 2022-01-13
> I'm personally in favour of having the default methods return a number and adding the convenience methods for string outputs, but the only thing I'm worried about if these changes will get caught, because if the TypeScript types are not checked, it could lead to silent misbehaviour (most likely with string concatenation), because there will not be any runtime error from that.
> There are other breaking changes and it's a new major version, so I guess it doesn't really matter. What do you think?

I completely agree. I think the risk of such a structural change going unnoticed on a major version update is quite minor, particularly if it's highlighted in the release notes.

As far as naming goes, how about...
```
h64(string[, BigInt]): BigInt
h64Raw(UInt8Array[, BigInt]: BigInt
h64Hex(string[,BigInt]): string
```