@jungomi 2022-01-14
`h64Hex` I don't associate Hex with string. I think it's easiest to keep it close to any string conversion naming, either `h64ToString` to suggest the `.toString()`, or maybe `h64AsString`. The fact that it's converted to a zero padded hex representation is the implementation detail, which is what would be expected from a `.toString()` if there were a native hex number type.
I also thought your current implementation with `h64String` was fine, just having them the other way around to match what would be my understanding of the names.