@marcusdarmstrong 2022-01-14
Works for me, I'll go with h64ToString().