@jungomi 2022-01-14
Thank you very much for your contributions. I'll be looking to prepare the release so that `v1.0.0` will be published in the next few days.