@extremeheat 2022-02-07
Regarding the switch from two numbers to BigInt for 64-bit seeds: my understanding is that BigInts are [much slower for almost all operations](https://github.com/tc39/proposal-bigint/issues/117) done on them--so is it safe to assume that the performance impact here for call overhead has been tested to be negligible?