@marcusdarmstrong 2022-02-07
In short, yes. I did detailed performance evaluation for every change included in this PR—in practice, there’s very little in the way of actual operations that need to be performed on a bigint in the relevant code paths: the math is all done on i64 types inside of easy and then boxed to a bigint on the way out. The only actual bigint operation is the u64-izing, which, while it has overhead, is significantly cheaper than the hi/low munging previously in use.